Refactor process_save_session_info_pdu for different infoData
The save session info PDU can contain several different kinds of data
structures in the infoData field. The old code didn't do a great job
of making this obvious, so I've substituted the if for a switch.

I extracted the implementation of the only supported logon info
structure to it's own function to make the code easier to read.

Finally, I added a warning when the server sends infoData we haven't
implemented support for.